Interworking with 2.6.18 kernel enabled
Notable is that to enable it for yaffs1 compatibility mode, I introduced
a workaround (translate_spare2oob/translate_oob2spare) which should be
redesigned if possible. The thing is that yaffs1 presumes SmartMedia OOB
layout which might be not true for a lot of modern flash controllers, so
there has to be translation.
Or, in case we get to tags inband, this workaround will become useless.