yaffs Yaffs browser can now create and delete hard links.
Defined values such as S_IFMT have now been added to yaffsfs.py

There is a bug with st.st_mode returning 0 if the object scanned is a symlink.
This is currently been worked on.
The ability to open and see symlinks in the browser is not working because of this.

Signed-off-by: Timothy Manning <tfhmanning@gmail.com>