yaffs: Refactor bit counting code
For Linux kernel we use hweight8() and hweight32(). This allows
use of special bit-counting code instructions in processors that
do this.

The yaffs direct we now only use one set of code. There were two.

NB On x86 the yaffs hweight8 code is actually a lot faster
than the sofwtare implementation in Linux kernel.

Signed-off-by: Charles Manning <cdhmanning@gmail.com>